examples/pyproject.toml
[tool.poetry]
name = "metawear.examples"

version = "0.1.0"
description = "Examples for the MetaWear-SDK-Python"
authors = ["MbientLab <hello@mbientlab.com>"]
repository = "https://github.com/nqminds/MetaWear-SDK-Python.git"
homepage = "https://github.com/nqminds/MetaWear-SDK-Python"

[tool.poetry.dependencies]
python = "^3.7"
metawear = { path = "../", develop = true }
bleak = "^0.14.3"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
